ajeitando o rest do cliente pra recer o id e criando o rest do histórico

+++ app/src/main/java/br/unifor/retail/rest/ProductService.java
@@ -21,5 +21,5 @@ public interface ProductService{
     @Headers({
             @Header(name = "X-Admin-Email", value = "admin@admin.com"),
             @Header(name = "X-Admin-Token", value = "C5TqmVb2GdaQJsPgy3mR")})
-    ResponseProduct searchProduct (@Path int idProduct);
+    ResponseProduct searchProduct (@Path Long idProduct);
 }
